+++ mitra_models/src/users/mod.rs
@@ -1,4 +1,3 @@
-pub mod helpers;
 pub mod queries;
 pub mod types;
 mod utils;