+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Set minimum migration version to 72.
+- Disabled custom migration that generates Ed25519 keys for accounts.
 
 ## [3.23.0] - 2025-04-15